--- test/runtime/gbt/tasks/callbacks/tcb-c.h
@@ -1,3 +1,0 @@
-void tcb_install_callbacks(void);
-void tcb_wait_for_nCallbacks(int);
-void tcb_report(void);
