--- test/runtime/gbt/tasks/callbacks/tcb-unin-c.h
@@ -1,4 +1,0 @@
-void tcb_install_callbacks(void);
-void tcb_wait_for_nCallbacks(int);
-void tcb_uninstall_one_callback(void);
-void tcb_report(void);
