+++ test/runtime/gbt/tasks/callbacks/tcb-unin.chpl
@@ -1,12 +1,14 @@
 proc main {
-  extern proc tcb_install_callbacks();
-  extern proc tcb_uninstall_one_callback();
+  extern proc tcb_install_callbacks_1();
+  extern proc tcb_install_callbacks_2();
+  extern proc tcb_uninstall_callbacks_1();
   extern proc tcb_wait_for_nCallbacks(nCallbacks: c_int);
   extern proc tcb_report();
 
   var counter: sync int = 0;
 
-  tcb_install_callbacks();
+  tcb_install_callbacks_1();
+  tcb_install_callbacks_2();
 
   cobegin {
     counter = counter + 1;
@@ -18,7 +20,7 @@ proc main {
   }
 
   tcb_wait_for_nCallbacks((if numLocales == 1 then 12 else 24): c_int);
-  tcb_uninstall_one_callback();
+  tcb_uninstall_callbacks_1();
 
   cobegin {
     counter = counter + 1;
