--- test/runtime/gbt/tasks/callbacks/tcb-unin.cleanfiles
@@ -1,1 +1,0 @@
-tcb-unin-c.o
