+++ test/runtime/gbt/tasks/callbacks/tcb-unin.compopts
@@ -1,1 +1,1 @@
-tcb-unin-c.h tcb-unin-c.o
+tcb-unin.h
