+++ test/runtime/gbt/tasks/callbacks/tcb-unin.h
@@ -1,0 +1,2 @@
+#define TCB_DATA_N 36 // only need 18 when #locales==1, but 36 otherwise
+#include "tcb-util.h"
