+++ test/runtime/gbt/tasks/callbacks/tcb-util.h
@@ -1,7 +1,29 @@
+//////////////////////
+//
+// Interface
+//
+
+void tcb_install_callbacks_1(void);
+void tcb_install_callbacks_2(void);
+void tcb_uninstall_callbacks_1(void);
+void tcb_wait_for_nCallbacks(int nCallbacks);
+void tcb_report(void);
+
+
+//////////////////////
+//
+// Implementation
+//
+
+#ifndef _tcb_util_h_
+#define _tcb_util_h_
+
 #include <pthread.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
+
+#include "chpl-tasks-callbacks.h"
 
 
 // This seems not to come from <string.h> without effort; just declare it.
@@ -18,8 +40,13 @@ static int tcb_data_idx = 0;
 static pthread_mutex_t tcb_data_mux = PTHREAD_MUTEX_INITIALIZER;
 
 
-void tcb_record(const char* prefix,
-                const chpl_task_cb_info_t* info) {
+////////////
+//
+// Private stuff
+//
+
+static void tcb_record(const char* prefix,
+                       const chpl_task_cb_info_t* info) {
   if (pthread_mutex_lock(&tcb_data_mux) != 0) {
     perror("pthread_mutex_lock()");
     exit(1);
@@ -59,6 +86,110 @@ void tcb_record(const char* prefix,
 
   if (pthread_mutex_unlock(&tcb_data_mux) != 0) {
     perror("pthread_muutex_lock()");
+    exit(1);
+  }
+}
+
+
+static void cb_create_1(const chpl_task_cb_info_t* info) {
+  tcb_record("CB1", info);
+}
+
+
+static void cb_begin_1(const chpl_task_cb_info_t* info) {
+  tcb_record("CB1", info);
+}
+
+
+static void cb_end_1(const chpl_task_cb_info_t* info) {
+  tcb_record("CB1", info);
+}
+
+
+static void cb_any_2(const chpl_task_cb_info_t* info) {
+  tcb_record("CB2", info);
+}
+
+
+////////////
+//
+// Public stuff
+//
+
+void tcb_install_callbacks_1(void) {
+  if (chpl_task_install_callback(chpl_task_cb_event_kind_create,
+                                 chpl_task_cb_info_kind_full,
+                                 cb_create_1)
+      != 0) {
+    fprintf(stderr, "Cannot install cb_create_1!\n");
+    exit(1);
+  }
+
+  if (chpl_task_install_callback(chpl_task_cb_event_kind_begin,
+                                 chpl_task_cb_info_kind_full,
+                                 cb_begin_1)
+      != 0) {
+    fprintf(stderr, "Cannot install cb_begin_1!\n");
+    exit(1);
+  }
+
+  if (chpl_task_install_callback(chpl_task_cb_event_kind_end,
+                                 chpl_task_cb_info_kind_id_only,
+                                 cb_end_1)
+      != 0) {
+    fprintf(stderr, "Cannot install cb_end_1!\n");
+    exit(1);
+  }
+}
+
+
+void tcb_install_callbacks_2(void) {
+  if (chpl_task_install_callback(chpl_task_cb_event_kind_create,
+                                 chpl_task_cb_info_kind_full,
+                                 cb_any_2)
+      != 0) {
+    fprintf(stderr, "Cannot install cb_any_2 for create!\n");
+    exit(1);
+  }
+
+  if (chpl_task_install_callback(chpl_task_cb_event_kind_begin,
+                                 chpl_task_cb_info_kind_full,
+                                 cb_any_2)
+      != 0) {
+    fprintf(stderr, "Cannot install cb_any_2 for begin!\n");
+    exit(1);
+  }
+
+  if (chpl_task_install_callback(chpl_task_cb_event_kind_end,
+                                 chpl_task_cb_info_kind_id_only,
+                                 cb_any_2)
+      != 0) {
+    fprintf(stderr, "Cannot install cb_any_2 for end!\n");
+    exit(1);
+  }
+}
+
+
+void tcb_uninstall_callbacks_1(void) {
+  //
+  // Uninstall the first callback, thus forcing the tasking layer to
+  // compact the list.
+  //
+  if (chpl_task_uninstall_callback(chpl_task_cb_event_kind_create, cb_create_1)
+      != 0) {
+    fprintf(stderr, "Cannot uninstall cb_create_1!\n");
+    exit(1);
+  }
+
+  if (chpl_task_uninstall_callback(chpl_task_cb_event_kind_begin, cb_begin_1)
+      != 0) {
+    fprintf(stderr, "Cannot uninstall cb_begin_1!\n");
+    exit(1);
+  }
+
+  if (chpl_task_uninstall_callback(chpl_task_cb_event_kind_end, cb_end_1)
+      != 0) {
+    fprintf(stderr, "Cannot uninstall cb_end_1!\n");
     exit(1);
   }
 }
@@ -127,3 +258,5 @@ void tcb_report(void) {
     exit(1);
   }
 }
+
+#endif // _tcb_util_h_
