--- test/runtime/gbt/tasks/callbacks/tcb.cleanfiles
@@ -1,1 +1,0 @@
-tcb-c.o
