+++ test/runtime/gbt/tasks/callbacks/tcb.compopts
@@ -1,1 +1,1 @@
-tcb-c.h tcb-c.o
+tcb.h
