+++ test/runtime/gbt/tasks/callbacks/tcb.h
@@ -1,0 +1,2 @@
+#define TCB_DATA_N 24 // only need 12 when #locales==1, but 24 otherwise
+#include "tcb-util.h"
